requirements.txt
selenium == 4.5.0
pytest == 7.2.0
webdriver-manager == 3.8.4
allure-pytest == 2.11.1